ref(StructuredData): Move some helper components out of StructuredData and isolate internals
